requirements.txt
Django==5.1.4
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
django-liveview==2.1.7
redis==5.2.1
django-unicorn==0.62.0
django-htmx==1.22.0
django-reactor==5.3.0b0
